# swagger/receipts-api.json
{"openapi":"3.0.0","paths":{"/":{"get":{"description":"Simple health endpoint to verify that the Receipts Validation API is running.","operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"summary":"Health check","tags":["health"]}},"/receipts":{"post":{"description":"Accepts a file and metadata. For milestone 1, only stores data in memory and does not perform OCR or validation.","operationId":"ReceiptsController_uploadReceipt","parameters":[],"requestBody":{"required":true,"description":"Receipt upload payload","content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UploadReceiptDto"}}}},"responses":{"201":{"description":"Receipt stored in memory (mock).","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReceiptDetailsDto"}}}}},"summary":"Upload a receipt with metadata (mock implementation).","tags":["receipts"]},"get":{"description":"Returns a summary list of all uploaded receipts stored in memory.","operationId":"ReceiptsController_listReceipts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ReceiptSummaryDto"}}}}}},"summary":"List all receipts (mock).","tags":["receipts"]}},"/receipts/{id}":{"get":{"description":"Returns details of a single receipt stored in memory. Throws 404 if not found.","operationId":"ReceiptsController_getReceipt","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReceiptDetailsDto"}}}},"404":{"description":"Receipt not found."}},"summary":"Get a single receipt by id (mock).","tags":["receipts"]}},"/receipts/validate":{"post":{"description":"For milestone 1, this endpoint always returns \"not_implemented\". In later milestones, it will validate metadata against extracted data.","operationId":"ReceiptsController_validateMetadataOnly","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReceiptDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationResultDto"}}}}},"summary":"Validate metadata only (no id, mock).","tags":["receipts"]}},"/receipts/{id}/validate":{"post":{"description":"For milestone 1, this endpoint checks that the receipt exists, but still returns \"not_implemented\".","operationId":"ReceiptsController_validateReceiptById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReceiptDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidationResultDto"}}}},"404":{"description":"Receipt not found."}},"summary":"Validate a specific receipt by id (mock).","tags":["receipts"]}}},"info":{"title":"Receipts Validation API","description":"API for uploading receipts and validating basic metadata. Current milestone: mock endpoints, no OCR, no business rules yet.","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UploadReceiptDto":{"type":"object","properties":{"amount":{"type":"number","example":100,"description":"Total amount of the receipt as provided by the user."},"date":{"type":"string","example":"2025-09-25","description":"Date of the event as provided by the user (ISO 8601)."},"event":{"type":"string","example":"Team lunch","description":"Business event for which the receipt was issued."},"file":{"type":"string","format":"binary","description":"Uploaded receipt file (PDF / JPG / PNG)."}},"required":["amount","date","event","file"]},"CreateReceiptDto":{"type":"object","properties":{"amount":{"type":"number","example":100,"description":"Total amount of the receipt as provided by the user."},"date":{"type":"string","example":"2025-09-25","description":"Date of the event as provided by the user (ISO 8601)."},"event":{"type":"string","example":"Team lunch","description":"Business event for which the receipt was issued."}},"required":["amount","date","event"]},"ReceiptDetailsDto":{"type":"object","properties":{"id":{"type":"string","example":"1","description":"Internal identifier of the receipt."},"originalFileName":{"type":"string","example":"receipt_123.pdf","description":"Original file name of the uploaded receipt."},"metadata":{"description":"Metadata provided by the user when uploading the receipt.","allOf":[{"$ref":"#/components/schemas/CreateReceiptDto"}]}},"required":["id","originalFileName","metadata"]},"ReceiptSummaryDto":{"type":"object","properties":{"id":{"type":"string","example":"1","description":"Internal identifier of the receipt."},"originalFileName":{"type":"string","example":"receipt_123.pdf","description":"Original file name of the uploaded receipt."},"amount":{"type":"number","example":100,"description":"Total amount provided in the metadata."},"date":{"type":"string","example":"2025-09-25","description":"Event date provided in the metadata."},"event":{"type":"string","example":"Team lunch","description":"Event description provided in the metadata."}},"required":["id","originalFileName","amount","date","event"]},"ValidationResultDto":{"type":"object","properties":{"status":{"type":"string","example":"not_implemented","description":"Current status of the validation logic."},"message":{"type":"string","example":"Validation logic is not implemented yet. OCR and business rules will be added in the next milestones.","description":"Human-readable explanation of the status."},"receiptId":{"type":"string","example":"1","description":"Optional receipt identifier that was requested for validation."}},"required":["status","message"]}}}}